Consolidated AttributeValueFacade and AttributeValueCollection
Decided that the AttributeValueFacade wasn't providing enough value since a) it still mandated access to the underlying collection, and b) the underlying collection doesn't provide much complexity (or risk of abuse).
